Add awslog mode
There is a bug in AWS SDK for some reason I can not run multiple awslog config
with multiple stream prefixes - the routine just go to null except the last
entries. Lots of debugging - shoing go routine is not the problems, seems it is
with aws sdk! esp sharing and session concurrancy.

This allow us to run multiple instance of the command for each config file
contains different awslog entry.