Fixed bugs allowing the use of multiple awslogs
This is go bug!. Or race conditions when using local variables.

The fix is not to pass the pointer to the function but pass by value copy. In
theory the existing code should work as it is, but it is not. When changing
pass by value copy things start working.