Implement command proxy mode.
The current server can work as a command proxy mode like below:

Some app wants to run a remote command on remote nsre instance but they only
provide a push URL without the ability to add JWT auth.

We take that url as localhost and local nsre take that path and send it to
remote nsre.

The proxy will handle the jwt authentication in its own config.