Limit the rows to 8
The solution of calculating rows is not working well <sigh> but no work around
at the moment. The rows keep increasing (probably the unmatch of the pixel and
the row height. In firefox it is even worse for one row, In chrome one row
retains it well.

For now just limit them to 8